feat(resamp): the control port is real, and observable — closes #670
Both halves of the control port's Python face, in one pass because
fixing either alone leaves it half-done: a double[] control nobody can
observe, or an observable mu fed through a complex port.

F2 — the block port took `const float _Complex *ctrl` while its own
streaming twin took a `double`, so half of every ctrl array was read and
discarded with no error and no warning. Now `double` on both faces,
matching nco and lo, which had the same narrowing fixed and whose
manifests already declare `double[]`. The production caller shows what it
cost: doppler_channel computed its Doppler dilation in double and cast it
down to reach this port. That cast is gone.

F1 — `resamp_get_ctrl_acc` is, per its own header, "the only way to see
what a closed timing loop is actually doing to the sampling instant", and
Resampler did not expose it. Now the read-only `Resampler.ctrl_acc`.
Verified from Python: 0.0 fresh, 0.0396 after a 0.01 steer, and 0.0 for a
caller driving the object through execute() -- which the docstring states
outright, because that path advances a different accumulator and a silent
permanent zero is exactly the kind of thing this campaign exists to
surface.

BREAKING, on the Python face only: a `complex64` ctrl array is now a
TypeError instead of a silent truncation, since numpy will not safe-cast
complex to float. Everything real still works -- float64, float32 and a
plain list are all accepted and asserted as a limit. This is the intended
consequence rather than collateral: discarding an imaginary part without
telling anyone was the defect.

§19 is retired, not deleted-and-forgotten. It pinned "only the real part
of ctrl is used"; a `double` port has no imaginary half, so the claim it
guarded no longer exists. The retirement is recorded in place so a later
reader does not read the gap as coverage that went missing.

Two things worth knowing for the next fragment change:

  * jm injects a declaration but does NOT rewrite an existing one whose
    signature changed -- Resampler_core.h still carried the complex
    prototype after `jm apply` and had to be corrected by hand.
  * deleting a sacred fragment to regenerate it is only safe if it is
    genuinely generated, and "carries no jm:hand marker" is NOT that
    test. resample_ext_Resampler.c had a hand-written `bank=` ctor path
    with no marker; regenerating dropped it and took out two construction
    tests. Restored from git and patched by hand instead.

resamp's report: 9 findings, ZERO open gaps, 14/14 limits. Verified with
ctest 100/100, 336 Python tests across resample + state serialization,
test-stubs 28, drift-check OK, validate-check OK, doxygen 0 warnings at
CI's 1.9.8.

+++ src/doppler/resample/tests/test_resampler.py
@@ -216,14 +216,14 @@ class TestExecuteCtrl:
     def test_unity_rate_zero_ctrl_count(self):
         r = Resampler(1.0)
         x = _ones(64)
-        ctrl = np.zeros(64, dtype=np.complex64)
+        ctrl = np.zeros(64, dtype=np.float64)
         y = r.execute_ctrl(x, ctrl)
         assert len(y) == 64
 
     def test_ctrl_too_short_raises(self):
         r = Resampler(1.0)
         x = _ones(64)
-        ctrl = np.zeros(32, dtype=np.complex64)
+        ctrl = np.zeros(32, dtype=np.float64)
         with pytest.raises(ValueError):
             r.execute_ctrl(x, ctrl)
 
@@ -231,7 +231,7 @@ class TestExecuteCtrl:
         r_base = Resampler(1.0)
         r_ctrl = Resampler(1.0)
         x = _ones(128)
-        ctrl = np.full(128, 0.5, dtype=np.complex64)
+        ctrl = np.full(128, 0.5, dtype=np.float64)
         y_base = r_base.execute(x)
         y_ctrl = r_ctrl.execute_ctrl(x, ctrl)
         assert len(y_ctrl) > len(y_base)
@@ -246,7 +246,7 @@ class TestExecuteCtrl:
     def test_decimation_with_ctrl(self):
         r = Resampler(0.5)
         x = _ones(128)
-        ctrl = np.zeros(128, dtype=np.complex64)
+        ctrl = np.zeros(128, dtype=np.float64)
         y = r.execute_ctrl(x, ctrl)
         assert 56 <= len(y) <= 64
 
@@ -267,7 +267,7 @@ class TestOutParam:
     def test_execute_ctrl_out_writes_into_callers_buffer(self):
         r = Resampler(1.0)
         x = _ones(64)
-        ctrl = np.zeros(64, dtype=np.complex64)
+        ctrl = np.zeros(64, dtype=np.float64)
         out = np.zeros(
             max(r.execute_ctrl_max_out(), len(x)), dtype=np.complex64
         )
@@ -277,7 +277,7 @@ class TestOutParam:
     def test_execute_ctrl_out_undersized_raises(self):
         r = Resampler(1.0)
         x = _ones(64)
-        ctrl = np.zeros(64, dtype=np.complex64)
+        ctrl = np.zeros(64, dtype=np.float64)
         with pytest.raises(ValueError):
             r.execute_ctrl(x, ctrl, out=np.zeros(1, dtype=np.complex64))
 
@@ -311,7 +311,7 @@ class TestOutParam:
     def test_execute_ctrl_out_unconvertible_raises(self):
         r = Resampler(1.0)
         x = _ones(64)
-        ctrl = np.zeros(64, dtype=np.complex64)
+        ctrl = np.zeros(64, dtype=np.float64)
         with pytest.raises((TypeError, ValueError)):
             r.execute_ctrl(x, ctrl, out="not an array")
